# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workflow/config": {
      "get": {
        "operationId": "WorkflowController_getConfig",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow configs retrieved"
          }
        },
        "summary": "Get service-level workflow configurations",
        "tags": [
          "workflow"
        ]
      },
      "patch": {
        "operationId": "WorkflowController_updateConfig",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow config updated"
          }
        },
        "summary": "Update a service workflow configuration",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/workflow/transition": {
      "post": {
        "operationId": "WorkflowController_transition",
        "parameters": [
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transition successful"
          }
        },
        "summary": "Transition application status",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/workflow/history/{appId}": {
      "get": {
        "operationId": "WorkflowController_getHistory",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History retrieved"
          }
        },
        "summary": "Get application transition history",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/workflow/audit-logs": {
      "get": {
        "operationId": "WorkflowController_getAuditLogs",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs retrieved"
          }
        },
        "summary": "Get system audit logs",
        "tags": [
          "workflow"
        ]
      }
    },
...
  },
  "info": {
    "title": "UCSDP Backend API",
    "description": "Unified Citizen Service and Delivery Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TransitionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}